+++ src/components/sections/About.astro
@@ -4,6 +4,7 @@ import Card from '@/components/ui/Card.astro';
 import TerminalHeader from '@/components/modules/TerminalHeader.astro';
 import StatCard from '@/components/ui/StatCard.astro';
 import ChipTech from '@/components/ui/ChipTech.astro';
+import SectionTitle from '@/components/ui/SectionTitle.astro';
 import {
   IconBrandReact,
   IconBrandAstro,
@@ -24,6 +25,8 @@ import {
 
 <section id="about" class="py-24 md:py-32">
   <div class="container-narrow about-section-content space-y-8">
+    <SectionTitle title="Sobre Mí" subtitle="Desarrollador Full Stack apasionado por crear experiencias digitales excepcionales." />
+
     <Card padding="none" class="overflow-hidden">
       <TerminalHeader />
     </Card>
