+++ src/components/sections/Certifications.astro
@@ -2,6 +2,7 @@
 // src/sections/Certifications.astro
 import { getCollection } from 'astro:content';
 import CertificationCard from '@/components/modules/CertificationCard.astro';
+import SectionTitle from '@/components/ui/SectionTitle.astro';
 
 const allCertifications = await getCollection('certifications');
 allCertifications.sort(
@@ -10,10 +11,8 @@ allCertifications.sort(
 ---
 
 <section id="certifications" class="py-24 md:py-32">
-  <div class="container-narrow certifications-section-content">
-    <h2 class="mb-16 text-center text-h2 font-bold text-text-primary">
-      // Certificaciones
-    </h2>
+  <div class="container-narrow certifications-section-content space-y-16">
+    <SectionTitle title="Certificaciones" subtitle="Credenciales técnicas que validan mi expertise." />
 
     <div class="grid grid-cols-1 gap-8 md:grid-cols-2">
       {
@@ -50,7 +49,6 @@ allCertifications.sort(
           ease: 'power3.out',
         });
       },
-      once: true,
     });
   });
 </script>
