+++ src/components/sections/Contact.astro
@@ -7,6 +7,7 @@ import Card from '@/components/ui/Card.astro';
 import ContactForm from '@/components/modules/ContactForm.astro';
 import Button from '@/components/ui/Button.astro';
 import StatusChip from '@/components/ui/StatusChip.astro';
+import SectionTitle from '@/components/ui/SectionTitle.astro';
 import {
   IconBrandGithub,
   IconBrandLinkedin,
@@ -18,17 +19,11 @@ import {
 ---
 
 <section id="contact" class="py-24 md:py-32">
-  <div class="container-narrow contact-section-content">
-    <div class="text-center">
-      <h2 class="text-h2 font-bold text-text-primary">// Contacto</h2>
-      <p
-        class="mx-auto mt-4 max-w-2xl text-lg text-text-secondary"
-        style="font-weight: 300;"
-      >
-        Colaboremos en tu próximo proyecto. Si tienes una pregunta, una
-        propuesta o simplemente quieres saludar, me encantaría saber de ti.
-      </p>
-    </div>
+  <div class="container-narrow contact-section-content space-y-16">
+    <SectionTitle
+      title="Contacto"
+      subtitle="Colaboremos en tu próximo proyecto. Si tienes una pregunta, una propuesta o simplemente quieres saludar, me encantaría saber de ti."
+    />
 
     <div class="mt-16 grid grid-cols-1 gap-8 md:grid-cols-2 md:items-start">
       <!-- COLUMNA IZQUIERDA: INFORMACIÓN Y ESTADO -->
@@ -106,7 +101,7 @@ import {
 
   gsap.registerPlugin(ScrollTrigger);
 
-  document.addEventListener('DOMContentLoaded', () => {
+  function initContactAnimation() {
     const section = document.querySelector('.contact-section-content');
     if (!section) return;
 
@@ -123,11 +118,13 @@ import {
           autoAlpha: 1,
           y: 0,
           duration: 1,
-          stagger: 0.15, // Una cascada más rápida y sutil
+          stagger: 0.15,
           ease: 'power3.out',
         });
       },
-      once: true,
     });
-  });
+  }
+
+  document.addEventListener('astro:page-load', initContactAnimation);
+  document.addEventListener('DOMContentLoaded', initContactAnimation);
 </script>
