+++ src/components/sections/Projects.astro
@@ -15,7 +15,7 @@ allProjects.sort(
     <!-- CENTRADO -->
     <div class="flex w-full justify-center">
       <SectionTitle
-        title="// Proyectos"
+        title="Proyectos"
         subtitle="Una selección de casos de estudio técnicos y creativos."
         class="items-center text-center"
       />
