+++ src/components/ui/Links.astro
@@ -21,7 +21,7 @@ const { ulClass, aClass } = Astro.props;
         <a
           href={link.href}
           class:list={[
-            'font-sans text-sm font-light uppercase tracking-widest text-text-secondary transition-colors hover:text-text-primary focus:outline-none focus-visible:ring-2 focus-visible:ring-text-primary',
+            'font-sans text-sm font-normal uppercase tracking-normal text-text-secondary transition-colors hover:text-[--color-accent-magenta] focus:outline-none focus-visible:ring-2 focus-visible:ring-text-primary',
             aClass,
           ]}
         >
